requirements.txt
aws-cdk-lib==2.102.0
aws-cdk.aws-cognito-identitypool-alpha==2.101.1a0
boto3==1.34.19
constructs>=10.0.0,<11.0.0
python-slugify==8.0.4